Refactored init function, updated tests.
This commit refactors the init() function to use 3 new private instance
methods (_getDeviceMotionController, _getDeviceOrientationController,
_getGravityCoefficient) to get the device motion/orientation controllers
and calculating the _gravityCoefficient value.

Chai-as-promised was removed from the dependencies as it doesn't work in
PhantomJS.
Some tests were refactored with stubs to allow testing in
browsers/devices that don't have the device orientation/motion APIs
(which is basically everything except Chrome).