Update import and usage documentation sections
Given that there is a "How to add" section, it is redundant to add the files in a HTML snippet again at "How to use".

I included the module import usage and examples. 

I also change the "How to use" snippet, to a pure JS one, since it was already clear on the previous section how to add GyroNorm.